Update patch set 1
Patch Set 1:

> >  So I think we
 > > should replace our master branch with something like [2]. Feel
 > free
 > > to handle as you wish.
 > >
 > > [1] http://pkgs.fedoraproject.org/cgit/rpms/collectd.git/commit/?id=ca83d8927dfe04f15458312633ed887c96ec7463
 > > [2] https://github.com/didib/collectd/commits/restored-history
 > 
 > We can not replace the branch with another branch. pushing
 > something directly to github will destroy synchronization etc.

I did not suggest to push to github. I pushed it there just to show what I mean. I suggested to push to gerrit. You can remove and add branches in gerrit, and while it's a bit ugly to maintain, it does work. So you can create a new "Old_master" from current master, then remove current master, and push (as a maintainer, skipping review) whatever you want to a new master. In fact I did that, and never thought about the sync to github. Now looked and it seems to be mostly ok. And the reason for that was exactly my current reason - to re-introduce the project's history. See, if you care, the comments on [1] leading to deciding to do this, and the resultant sync in github [2]. Indeed, when you eventually have a look at [3], you can't tell which patches went to the now-dead master and which were to the current, new master. I think I can live with that, if the gain is having the entire git history. But that's obviously up to you, as a maintainer.

[1] https://gerrit.ovirt.org/#/c/34736/
[2] https://github.com/oVirt/ovirt-wgt/branches
[3] https://gerrit.ovirt.org/#/q/project:ovirt-wgt

Patch-set: 1