Update patch set 3
Patch Set 3:

> I would prefer to keep config of a plugin and the load directive in
 > one file.

The config is currently commented out.

Not sure what exactly you want.

Copy the commented-out config to default-plugins.conf ?

Which parts? E.g. there is a '<Plugin cpu>' stanza, and there is an example for 'aggregation' that uses it. Both "<Plugin cpu>" and "<Plugin aggregation>" examples are separate, in different places, with (perhaps intended) different contexts - in default-plugins.conf they will not have that context.

Do this statically? This carries a higher change for future divergence between the copies.

Do this automatically? Not sure how to do this in a simple way, and not sure complex ways are worth it.

Remove the original texts? This carries a higher chance for future patch conflicts.

Patch-set: 3