Update patch set 3
Patch Set 3:

Was this meant to pick up the changes so that it matches the new version pending merge upstream? If not, would you like me to prepare that so it can be submitted for review if/when it is accepted?

Patch-set: 3
Reviewer: Gerrit User 567 <567@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0