Update patch set 7
Patch Set 7:

I think know why this is failing:
The package is misnamed when building on centos9. (.el. instead of .el9)

The failing task is defined in [1]

It calls gating_jobs/pkg_validate.sh
This calls the build_srpm function from rpm-koji-gating-lib.common [2]

There's a conditional for installing the openstack-macros.
It seems like these are already installed on centos-9 (probably because the preceeding task ran the pkg_validate script already).
Because the package is installed, L101 and 102 are not run, so DIST_VER is not set, and it is needed later, to set el9 as part of the package version. 

A quick fix might be to make sure to uninstall openstack macros in [1].

A better fix is to update [2]

[1] https://review.rdoproject.org/cgit/rdo-config/playbooks/opstools-perfmon-common-distgit-cbs-validate/run.yaml 
[2] https://review.rdoproject.org/cgit/gating_scripts/tree/rpm-koji-gating-lib.common#n101

Patch-set: 7